test_format/requirements.txt
avro==1.11.1
avro-python3==1.10.2
msgpack==1.0.5
protobuf==4.22.3
Pympler==1.0.1
pyyaml==5.4.1
